uapi: convert net/route.h
